refactor: separate segment importance logic
